:art: refactor(#182/auth): result dto 위치 이동 및 사용하지 않는 파일 삭제
