+++ ninedot-api/src/main/java/org/sopt36/ninedotserver/auth/v1/dto/response/SignupThenLoginResponse.java
@@ -1,6 +1,6 @@
 package org.sopt36.ninedotserver.auth.v1.dto.response;
 
-import org.sopt36.ninedotserver.auth.dto.response.SignupThenLoginResult;
+import org.sopt36.ninedotserver.auth.dto.result.SignupThenLoginResult;
 
 public record SignupThenLoginResponse(
     String accessToken,
