+++ ninedot-api/src/main/java/org/sopt36/ninedotserver/auth/v1/dto/response/UserRetrievedResponse.java
@@ -1,6 +1,6 @@
 package org.sopt36.ninedotserver.auth.v1.dto.response;
 
-import org.sopt36.ninedotserver.auth.dto.response.UserRetrievedResult;
+import org.sopt36.ninedotserver.auth.dto.result.UserRetrievedResult;
 
 public record UserRetrievedResponse(
     Long id,
