--- ninedot-application/src/main/java/org/sopt36/ninedotserver/auth/dto/response/NewAccessTokenResult.java
@@ -1,8 +1,0 @@
-package org.sopt36.ninedotserver.auth.dto.response;
-
-public record NewAccessTokenResult(
-    String accessToken,
-    String message
-) {
-
-}
