+++ ninedot-application/src/main/java/org/sopt36/ninedotserver/auth/port/in/SignupUsecase.java
@@ -1,7 +1,7 @@
 package org.sopt36.ninedotserver.auth.port.in;
 
 import org.sopt36.ninedotserver.auth.dto.command.SignupCommand;
-import org.sopt36.ninedotserver.auth.dto.response.SignupThenLoginResult;
+import org.sopt36.ninedotserver.auth.dto.result.SignupThenLoginResult;
 
 public interface SignupUsecase {
